@@ -5,12 +5,12 @@ version: 3.0.0
 homepage: https://github.com/De-Morgan/flutter-contextual-action-bar
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.17.6 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  provider: ^5.0.0
+  provider: ^6.0.3
 
 
 dev_dependencies:
