@@ -1,5 +1,5 @@
 export 'contextual_action.dart';
 
-export 'contextual_action_bar.dart';
+export 'contextual_app_bar.dart';
 
 export 'contextual_action_widget.dart';
